Refactoring the app terminal to reflect the Model-View-Controller schematic
